Move invalid_upcast_comparisons to its own module
